requirements.txt
Flask==2.3.4
Flask-SQLAlchemy==3.0.3
gunicorn==20.1.0
python-dotenv==1.0.0
